separate serde check
